agentcore/runtime/bulk-analysis-agent/requirements.txt
bedrock-agentcore>=0.1.7
strands-agents>=1.11.0
